quick change to fix for BUG 3091.

+++ src/main/java/ca/sqlpower/dao/SPPersister.java
@@ -38,7 +38,7 @@ public interface SPPersister {
 		INTEGER("Integer", Integer.class),
 		DOUBLE("Double", Double.class),
 		BOOLEAN("Boolean", Boolean.class),
-		SHORT("Long", Short.class),
+		SHORT("Short", Short.class),
 		LONG("Long", Long.class),
 		
 		/**
